2024-10-10 paskaitos kodas

+++ JavaScript/Eshop mini/deleteProduct.js
@@ -1,6 +1,10 @@
 function deleteProduct(index) {
-	// Iš masyvo šalinamas 1 elementas pradedant nuo index pozicijos
-	produktai.splice(index, 1);
-	updateHtmlTable();
-	localStorage.setItem("produktai", JSON.stringify(produktai));
+	const confirmation = confirm("Ar tikrai norite pašalinti šį produktą?");
+	//true/false
+	if (confirmation) {
+		// Iš masyvo šalinamas 1 elementas pradedant nuo index pozicijos
+		produktai.splice(index, 1);
+		updateHtmlTable();
+		localStorage.setItem("produktai", JSON.stringify(produktai));
+	}
 }
